yolo_server/requirements.txt
flask==3.0.2
flask-cors==4.0.0
ultralytics==8.1.28
numpy==1.26.4
Pillow>=10.0.0
python-dotenv==1.0.1 